+++ tests/data/TestProject/Code/Scenarios/defaultScenario.R
@@ -4,6 +4,9 @@ defaultScenario <- function(projectConfiguration, loadPreSimulatedResults = FALS
 
   # Define which scenarios to run
   scenarioNames <- c("TestScenario")
+  # Set scenario names to NULL if you want to simulate all scenarios defined in the
+  # excel file
+  # scenarioNames <- NULL
 
   # Create `ScenarioConfiguration` objects from excel files
   scenarioConfigurations <- readScenarioConfigurationFromExcel(
@@ -15,10 +18,10 @@ defaultScenario <- function(projectConfiguration, loadPreSimulatedResults = FALS
   # E.g., enable setting of test parameters. If set to `TRUE`, parameters
   # defined in "InputCoode/TestParameters.R" will be applied
   scenarioConfigurations[[1]]$setTestParameters <- FALSE
-  # Set output paths for each scenario
-  for (scenarioConfiguration in scenarioConfigurations) {
-    scenarioConfiguration$outputPaths <- enumValues(OutputPaths)
-  }
+
+  # Disable check for negative values if required
+  scenarioConfigurations[[1]]$simulationRunOptions <- SimulationRunOptions$new()
+  scenarioConfigurations[[1]]$simulationRunOptions$checkForNegativeValues <- FALSE
 
 
   customParams <- NULL
@@ -35,7 +38,7 @@ defaultScenario <- function(projectConfiguration, loadPreSimulatedResults = FALS
     saveScenarioResults(simulatedScenarios, projectConfiguration)
   }
 
-  ########### Load observed data - data template v10########
+  ########### Load observed data########
   # Which sheets to load
   dataSheets <- "Laskin 1982.Group A"
   observedData <- esqlabsR::loadObservedData(
@@ -44,7 +47,17 @@ defaultScenario <- function(projectConfiguration, loadPreSimulatedResults = FALS
   )
 
   ########## Create figures########
+  # Output the names of loaded data sets to conveniently transfer them to the excel
+  # file for figure specification
+  # sort(names(observedData))
+  ########## Create figures########
+  plots <- createPlotsFromExcel(
+    simulatedScenarios = simulatedScenarios,
+    observedData = observedData,
+    projectConfiguration = projectConfiguration,
+    stopIfNotFound = TRUE
+  )
 
-  # Return simulated scenarios
-  return(simulatedScenarios)
+  # Return a list with simulated scenarios and created plots
+  return(list(simulatedScenarios = simulatedScenarios, plots = plots))
 }