+++ NEWS.md
@@ -3,7 +3,7 @@
 ### BREAKING CHANGES
 
 - `Scenarios` excel file gets additional columns `SteadyStateTime`, `SteadyStateTimeUnit`, 
-`PopulationId`.
+`PopulationId`, `OutputPathsIds`.
 - `readScenarioConfigurationFromExcel()` has a new signature and requires a list of 
 `scenarioNames` and a `ProjectConfiguration`. The output is a named list of `ScenarioConfiguration` 
 objects.
@@ -14,6 +14,16 @@ from the respective field of `ScenarioConfgiruation`
 - Enum `GraphicsDevices` has been removed.
 
 ### MAJOR CHANGES
+
+- Paths of model outputs are defined in the excel file `Scenarios.xlsx`. In the 
+sheet `OutputPaths`, create an entry for each output. The column `OutputPath` is the full 
+path to the output, while `OutputPathId` can be an identifier that conveniently allows 
+to select the correct output.
+In the `Scenarios` sheet, enter the IDs of all paths the outputs should be generated for,
+separated by a `,`, e.g. `Aciclovir_PVB, Aciclovir_fat_cell`. 
+
+If no outputs are specified, the outputs as defined in the simulation `.pkml` file
+will be produced.
 
 - `ScenarioConfiguration` gets a new field `populationId`, specifying the id of 
 population as defined in the `PopulationParameters.xlsx` file, sheet `Demographics`.